chore: refactor audit crates
