+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
   "crates/bfv-client",
   "crates/ciphernode-builder",
   "crates/cli",
+  "crates/committee-hash",
   "crates/compute-provider",
   "crates/config",
   "crates/console",
@@ -32,6 +33,7 @@ members = [
   "crates/safe",
   "crates/sdk",
   "crates/daemon-server",
+  "crates/slashing",
   "crates/sortition",
   "crates/support-scripts",
   "crates/sync",
@@ -87,6 +89,7 @@ e3-data = { version = "0.1.15", path = "./crates/data" }
 e3-request = { version = "0.1.15", path = "./crates/request" }
 e3-sdk = { version = "0.1.15", path = "./crates/sdk" }
 e3-cli = { version = "0.1.15", path = "./crates/cli" }
+e3-committee-hash = { version = "0.1.15", path = "./crates/committee-hash" }
 e3-entrypoint = { version = "0.1.15", path = "./crates/entrypoint" }
 e3-init = { version = "0.1.15", path = "./crates/init" }
 e3-events = { version = "0.1.15", path = "./crates/events" }
@@ -113,6 +116,7 @@ e3-trbfv = { version = "0.1.15", path = "./crates/trbfv" }
 e3-utils = { version = "0.1.15", path = "./crates/utils" }
 e3-utils-derive = { version = "0.1.15", path = "./crates/utils-derive" }
 e3-safe = { version = "0.1.15", path = "./crates/safe" }
+e3-slashing = { version = "0.1.15", path = "./crates/slashing" }
 e3-zk-prover = { version = "0.1.15", path = "./crates/zk-prover" }
 e3-zk-helpers = { version = "0.1.15", path = "./crates/zk-helpers" }
 e3-parity-matrix = { version = "0.1.15", path = "./crates/parity-matrix" }
