+++ crates/cli/src/config_setup.rs
@@ -1,8 +1,16 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 //
 // This file is provided WITHOUT ANY WARRANTY;
-// without even even the implied warranty of MERCHANTABILITY
+// without even the implied warranty of MERCHANTABILITY
 // or FITNESS FOR A PARTICULAR PURPOSE.
+
+//! Configuration bootstrapping for `enclave setup`.
+//!
+//! Generates an `enclave.config.yaml` from compiled-in contract deployment
+//! addresses. This is a **dev-bootstrap tool** — it couples to the Solidity
+//! build artifacts at compile time and is only invoked by the `enclave setup`
+//! CLI command. Production deployments should use a proper deployment artifact
+//! file instead of this compile-time approach.
 
 use alloy::primitives::Address;
 use anyhow::{anyhow, bail, Result};
@@ -13,12 +21,9 @@ use std::fs;
 use std::path::PathBuf;
 use tracing::instrument;
 
-// Import a built file:
-//   see /target/debug/enclave-xxxxxx/out/contract_deployments.rs
-//   also see build.rs
+// Generated by build.rs from packages/enclave-contracts/deployed_contracts.json
 include!(concat!(env!("OUT_DIR"), "/contract_deployments.rs"));
 
-// Get the ContractInfo object
 fn get_contract_info(name: &str) -> Result<&ContractInfo> {
     Ok(CONTRACT_DEPLOYMENTS
         .get(name)
@@ -39,7 +44,7 @@ pub fn validate_eth_address(address: &String) -> Result<()> {
 
 #[instrument(name = "app", skip_all)]
 pub fn execute(rpc_url: &str, config_dir: &PathBuf) -> Result<AppConfig> {
-    fs::create_dir_all(&config_dir)?;
+    fs::create_dir_all(config_dir)?;
 
     let config_path = config_dir.join("enclave.config.yaml");
 
@@ -76,7 +81,6 @@ chains:
 
     fs::write(config_path.clone(), config_content)?;
 
-    // Load with default location
     let config = load_config("_default", Some(config_path.display().to_string()), None)?;
 
     Ok(config)
@@ -99,7 +103,6 @@ mod tests {
         assert!(
             validate_eth_address(&"0x0000000000000000000000000000000000000000".to_string()).is_ok()
         );
-
         Ok(())
     }
 }
