+++ crates/config/src/lib.rs
@@ -9,10 +9,12 @@ pub mod chain_config;
 pub mod contract;
 pub mod load_config;
 pub mod paths_engine;
+pub mod program_config;
 pub mod rpc;
 pub mod validation;
 mod yaml;
 
 pub use app_config::*;
 pub use contract::*;
+pub use program_config::*;
 pub use rpc::*;
