+++ crates/events/src/commitment_link.rs
@@ -1,0 +1,67 @@
+// SPDX-License-Identifier: LGPL-3.0-only
+//
+// This file is provided WITHOUT ANY WARRANTY;
+// without even the implied warranty of MERCHANTABILITY
+// or FITNESS FOR A PARTICULAR PURPOSE.
+
+//! Cross-circuit commitment consistency types.
+//!
+//! Defines the [`CommitmentLink`] trait and supporting types used by the
+//! [`CommitmentConsistencyChecker`] in `e3-slashing`.
+
+use crate::ProofType;
+
+/// A 32-byte field element extracted from proof public signals.
+pub type FieldValue = [u8; 32];
+
+/// How source and target proofs relate and where faults are attributed.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum LinkScope {
+    /// Both proofs are generated by the same party (same Ethereum address).
+    SameParty,
+    /// The source proof is per-party while the target is from a different party.
+    CrossParty,
+    /// Each source proof claims a value that must exist among the set of
+    /// target proof outputs (from any party). Fault is attributed to the source.
+    SourceMustExistInTargets,
+}
+
+/// Defines a cross-circuit commitment consistency check.
+///
+/// Implementations extract commitment values from the public signals of a
+/// *source* proof type and verify they are consistent with the public signals
+/// of a *target* proof type.
+pub trait CommitmentLink: Send + Sync {
+    /// Human-readable name for logging.
+    fn name(&self) -> &'static str;
+
+    /// The proof type that *produces* the commitment value.
+    fn source_proof_type(&self) -> ProofType;
+
+    /// The proof type that *consumes* or must agree with the commitment value.
+    fn target_proof_type(&self) -> ProofType;
+
+    /// Relationship scope between source and target parties.
+    fn scope(&self) -> LinkScope;
+
+    /// Extract the commitment value(s) from the source proof's public signals.
+    fn extract_source_values(&self, public_signals: &[u8]) -> Vec<FieldValue>;
+
+    /// Return `true` when `source_values` are consistent with
+    /// `target_public_signals`, without party context.
+    fn check_signals(&self, _source_values: &[FieldValue], _target_public_signals: &[u8]) -> bool {
+        unimplemented!("override check_signals or check_consistency")
+    }
+
+    /// Return `true` when `source_values` are consistent with
+    /// `target_public_signals`, with party context.
+    fn check_consistency(
+        &self,
+        source_values: &[FieldValue],
+        target_public_signals: &[u8],
+        _src_party_id: u64,
+        _tgt_party_id: u64,
+    ) -> bool {
+        self.check_signals(source_values, target_public_signals)
+    }
+}
