+++ crates/zk-prover/Cargo.toml
@@ -18,12 +18,14 @@ bincode = "1.3.3"
 bn254_blackbox_solver = { git = "https://github.com/noir-lang/noir", tag = "v1.0.0-beta.16" }
 chrono = { workspace = true }
 directories = "5"
+e3-committee-hash.workspace = true
 e3-config.workspace = true
 e3-data.workspace = true
 e3-events.workspace = true
 e3-fhe-params.workspace = true
 e3-polynomial.workspace = true
 e3-request.workspace = true
+e3-slashing.workspace = true
 e3-utils.workspace = true
 e3-zk-helpers.workspace = true
 fhe.workspace = true
