+++ crates/zk-prover/src/actors/commitment_links/mod.rs
@@ -6,11 +6,8 @@
 
 //! Cross-circuit commitment consistency links.
 //!
-//! Each [`CommitmentLink`] defines a relationship between two proof types
-//! where a commitment value produced by one circuit must match a value
-//! consumed or produced by another circuit. The
-//! [`CommitmentConsistencyChecker`](super::commitment_consistency_checker::CommitmentConsistencyChecker)
-//! evaluates these links as verified proofs arrive.
+//! Concrete implementations of [`CommitmentLink`](e3_events::CommitmentLink)
+//! for each ZK proof pair. The trait and supporting types live in `e3-events`.
 
 pub mod c0_to_c3;
 pub mod c1_to_c2;
@@ -21,77 +18,9 @@ pub mod c4a_to_c6;
 pub mod c4b_to_c6;
 pub mod c6_to_c7;
 
-use e3_events::ProofType;
+// Re-export the canonical trait and types from e3-events.
+pub use e3_events::{CommitmentLink, FieldValue, LinkScope};
 use e3_fhe_params::BfvPreset;
-
-/// A 32-byte BN254 field element extracted from public signals.
-pub type FieldValue = [u8; 32];
-
-/// How source and target proofs relate and where faults are attributed.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum LinkScope {
-    /// Both proofs are generated by the same party (same Ethereum address).
-    /// One source entry is compared to one target entry per address.
-    SameParty,
-    /// The source proof is per-party while the target is from a different
-    /// party (e.g. per-node C1 vs aggregator C5). Each source's extracted
-    /// value must appear in at least one target's public signals.
-    /// Fault is attributed to the source if no target matches.
-    CrossParty,
-    /// Each source proof claims a value that must exist among the set of
-    /// target proof outputs (from any party). Fault is attributed to the
-    /// **source** when no target matches. Used when there are many source
-    /// proofs (e.g. N-1 C3 proofs per sender) each referencing a different
-    /// target (e.g. the recipient's C0 pk_commitment).
-    SourceMustExistInTargets,
-}
-
-/// Defines a cross-circuit commitment consistency check.
-///
-/// Implementations extract commitment values from the public signals of a
-/// *source* proof type and verify they are consistent with the public signals
-/// of a *target* proof type.
-pub trait CommitmentLink: Send + Sync {
-    /// Human-readable name for logging.
-    fn name(&self) -> &'static str;
-
-    /// The proof type that *produces* the commitment value.
-    fn source_proof_type(&self) -> ProofType;
-
-    /// The proof type that *consumes* or must agree with the commitment value.
-    fn target_proof_type(&self) -> ProofType;
-
-    /// Relationship scope between source and target parties.
-    fn scope(&self) -> LinkScope;
-
-    /// Extract the commitment value(s) from the source proof's public signals.
-    fn extract_source_values(&self, public_signals: &[u8]) -> Vec<FieldValue>;
-
-    /// Return `true` when `source_values` are consistent with
-    /// `target_public_signals`, without party context.
-    ///
-    /// Most links implement this. Links that need positional party information
-    /// should override [`check_consistency`] instead and leave this unimplemented.
-    fn check_signals(&self, _source_values: &[FieldValue], _target_public_signals: &[u8]) -> bool {
-        unimplemented!("override check_signals or check_consistency")
-    }
-
-    /// Return `true` when `source_values` are consistent with
-    /// `target_public_signals`.
-    ///
-    /// `src_party_id` and `tgt_party_id` are 0-based committee indices.
-    /// Defaults to ignoring party IDs and delegating to [`check_signals`].
-    /// Override this only when positional party information is required.
-    fn check_consistency(
-        &self,
-        source_values: &[FieldValue],
-        target_public_signals: &[u8],
-        _src_party_id: u64,
-        _tgt_party_id: u64,
-    ) -> bool {
-        self.check_signals(source_values, target_public_signals)
-    }
-}
 
 /// Returns the default set of commitment links to register.
 ///
